Move methods used for payu status to a separate file from telemetry
